rename: style 디렉터리명 수정 (style → styles)
