+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { CartProvider } from './contexts/CartContext';
 import Products from '@/pages/Products';
 
-import baseStyle from '@/style/base.style';
+import baseStyle from '@/styles/base.style';
 import { Global } from '@emotion/react';
 function App() {
   return (
